% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimNetR_functions.R
\name{find.knn}
\alias{find.knn}
\title{FUNCTION_TITLE}
\usage{
find.knn(embeddings, query, k)
}
\arguments{
\item{embeddings}{PARAM_DESCRIPTION}

\item{query}{PARAM_DESCRIPTION}

\item{k}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
