% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimNetR_functions.R
\name{network.dynamics.ODE.2}
\alias{network.dynamics.ODE.2}
\title{FUNCTION_TITLE}
\usage{
network.dynamics.ODE.2(
  net,
  state,
  mode = "fuzzy.logic",
  h = 50,
  gamma = 1,
  time.limit = 30,
  time.step.size = 0.01,
  weights,
  perturbations = NULL
)
}
\arguments{
\item{net}{PARAM_DESCRIPTION}

\item{state}{PARAM_DESCRIPTION}

\item{mode}{PARAM_DESCRIPTION, Default: 'fuzzy.logic'}

\item{h}{PARAM_DESCRIPTION, Default: 50}

\item{gamma}{PARAM_DESCRIPTION, Default: 1}

\item{time.limit}{PARAM_DESCRIPTION, Default: 30}

\item{time.step.size}{PARAM_DESCRIPTION, Default: 0.01}

\item{weights}{PARAM_DESCRIPTION}

\item{perturbations}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
