% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network}
\alias{network}
\title{SimNetR-Network Data}
\format{
A dataframe encoding the network in 5 columns:'Source','Interaction','Target','Weights' and 'String'; the latter includes the content of 'Source','Interaction' and 'Target' space separated.
}
\usage{
network
}
\description{
A dataset containing the transcriptional network inferred using SCENIC pipeline on murine TILs transcriptomics.
}
\examples{
network
}
\keyword{datasets}
