% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimNetR_functions.R
\name{perturbNodes}
\alias{perturbNodes}
\title{FUNCTION_TITLE}
\usage{
perturbNodes(nodes, at.times, duration, intensity, time.step.size, times)
}
\arguments{
\item{nodes}{PARAM_DESCRIPTION}

\item{at.times}{PARAM_DESCRIPTION}

\item{duration}{PARAM_DESCRIPTION}

\item{intensity}{PARAM_DESCRIPTION}

\item{time.step.size}{PARAM_DESCRIPTION}

\item{times}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
