% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phenotypes}
\alias{phenotypes}
\title{SimNetR-Phenotypes Data}
\format{
A dataframe containing network states, with nodes in raws and phenotypes in columns
}
\usage{
phenotypes
}
\description{
A dataset containing several network stable states
}
\examples{
phenotypes
}
\keyword{datasets}
