% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimNetR_functions.R
\name{plotPerturbation}
\alias{plotPerturbation}
\title{FUNCTION_TITLE}
\usage{
plotPerturbation(dat, col = NULL, title = "Node activity in time")
}
\arguments{
\item{dat}{PARAM_DESCRIPTION}

\item{col}{PARAM_DESCRIPTION, Default: NULL}

\item{title}{PARAM_DESCRIPTION, Default: 'Node activity in time'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
