% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimNetR_functions.R
\name{range01.w.max.min}
\alias{range01.w.max.min}
\title{FUNCTION_TITLE}
\usage{
range01.w.max.min(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
