% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimNetR_functions.R
\name{scale_vector}
\alias{scale_vector}
\title{FUNCTION_TITLE}
\usage{
scale_vector(v, lower_limit, upper_limit)
}
\arguments{
\item{v}{PARAM_DESCRIPTION}

\item{lower_limit}{PARAM_DESCRIPTION}

\item{upper_limit}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
