% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimNetR_functions.R
\name{strsplit2}
\alias{strsplit2}
\title{FUNCTION_TITLE}
\usage{
strsplit2(x, split, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{split}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
