% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimNetR_functions.R
\name{plotPerturbation}
\alias{plotPerturbation}
\title{FUNCTION_TITLE}
\usage{
plotPerturbation(data, title, xlabel, ylabel)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{title}{PARAM_DESCRIPTION}

\item{xlabel}{PARAM_DESCRIPTION}

\item{ylabel}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
